render fixes and controller update

+++ gameplay/object.go
@@ -1,6 +1,6 @@
 package gameplay
 
 type Object struct {
-	X, Y, Speed float32 
+	X, Y, Speed float64
 	Direction float64
 }
